+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 ## [v16.1.0-rc1.0.8] - 2022-03-28
 ### Changed
 - Support for configuring claims supported in Keycloak OP metadata [RCIAM-899](https://jira.argo.grnet.gr/browse/RCIAM-899)
+### Fixed
+- Include 'urn:ietf:params:oauth:grant-type:token-exchange' in grant_types_supported field of Keycloak OP metadata, if token-exchange is enabled [RCIAM-915](https://jira.argo.grnet.gr/browse/RCIAM-915)
 
 ## [v16.1.0-rc1.0.7] - 2022-03-17
 ### Changed